requirements.txt
# system requirements
# Usage: pip install -r requirements.txt

# Base ------------------------------------------------------------------------
matplotlib>=3.7.1
numpy>=1.22.4
kneed>=0.8.2
pandas>=1.4.4
pip>=22.0.4
threadpoolctl>=3.1.0
scikit-learn>=1.2.2
scipy>=1.10.1
sklearn-pandas>=2.2.0